Rate limit PyPI requests in run.py to comply with recent PyPI changes
